+++ src/Support/Mutation.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Support;
 
-class Mutation extends Field
+abstract class Mutation extends Field
 {
 }