+++ src/Support/Query.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Support;
 
-class Query extends Field
+abstract class Query extends Field
 {
 }